app/requirements.txt
Flask==2.3.3
psycopg2-binary>=2.9.1
Werkzeug==2.3.7